requirement.txt
flask_sqlalchemy==3.1.1
SQLite== 3.45.3 